+++ src/Commands/Refresh.js
@@ -49,11 +49,11 @@ class Refresh extends Command {
       this.checkEnv(flags.force)
       const migrationsFiles = this.loadFiles(this.helpers.migrationsPath())
       const MigrationsRunner = this.migrations
-      yield new MigrationsRunner().down(migrationsFiles, 0)
-      const response = yield new MigrationsRunner().up(migrationsFiles)
+      yield new MigrationsRunner().down(migrationsFiles, 0, flags.log)
+      const response = yield new MigrationsRunner().up(migrationsFiles, flags.log)
       const successMessage = 'Migrations successfully refreshed.'
       const infoMessage = 'Already at the latest batch.'
-      this.log(response.status, successMessage, infoMessage)
+      this._log(response.status, successMessage, infoMessage)
     } catch (e) {
       this.error(e)
     }